cast to int64

+++ pkg/storage/unified/sql/server.go
@@ -59,8 +59,8 @@ func NewResourceServer(ctx context.Context, db infraDB.DB, cfg *setting.Cfg,
 		opts.Search = resource.SearchOptions{
 			Backend: search.NewBleveBackend(search.BleveOptions{
 				Root:          cfg.IndexPath,
-				FileThreshold: cfg.IndexFileThreshold, // fewer than X items will use a memory index
-				BatchSize:     cfg.IndexMaxBatchSize,  // This is the batch size for how many objects to add to the index at once
+				FileThreshold: int64(cfg.IndexFileThreshold), // fewer than X items will use a memory index
+				BatchSize:     cfg.IndexMaxBatchSize,         // This is the batch size for how many objects to add to the index at once
 			}, tracer, reg),
 			Resources:     docs,
 			WorkerThreads: cfg.IndexWorkers,
